+++ intellij/play/src/CustomCommand.java
@@ -37,21 +37,18 @@ public class CustomCommand extends Entity {
     }
 
     public String generateSaveString() {
-        String str = "" + name + "\n" + description + "\n" + uid + "\n" + syntaxRegex + "\n" + parameters;
+        StringBuilder str = new StringBuilder("" + name + "\n" + description + "\n" + uid + "\n" + syntaxRegex + "\n" + parameters);
         for (int i = 0; i < eventName.size(); i++)
-            str = str + "\n++ev++" + eventName.get(i) + "---" + eventCode.get(i);
-        for (int i = 0; i < tags.size(); i++)
-            str = str + "\n++tag++" + tags.get(i);
+            str.append("\n++ev++").append(eventName.get(i)).append("---").append(eventCode.get(i));
+        for (String tag : tags) str.append("\n++tag++").append(tag);
         for (int i = 0; i < localVarName.size(); i++)
-            str = str + "\n++variable++" + localVarUids.get(i) + "---" + localVarName.get(i) + "---" + localVarType.get(i) + "---" + localVarValue.get(i) + "\n";
-        return str;
+            str.append("\n").append("\n++variable++").append(localVarUids.get(i)).append("---").append(localVarName.get(i)).append("---").append(localVarType.get(i)).append("---").append(localVarValue.get(i));
+        return str.toString();
     }
 
     public boolean matches(String command) {
         Matcher m = patt.matcher(command);
-        if (m.find())
-            return true;
-        return false;
+        return m.find();
     }
 
     public String[] getArgs(String command) {